requirements.txt
playsound
Pillow